Refactor: Adapter Package Structure Change
