+++ src/main/kotlin/com/example/ggyunispring/adapter/ResponseDTO.kt
@@ -1,4 +1,4 @@
-package com.example.ggyunispring.adapter.dto.response
+package com.example.ggyunispring.adapter
 
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
