+++ src/main/kotlin/com/example/ggyunispring/application/DiaryService.kt
@@ -2,9 +2,9 @@ package com.example.ggyunispring.application
 
 import com.example.ggyunispring.domain.diary.Diary
 import com.example.ggyunispring.domain.diary.DiaryRepository
-import com.example.ggyunispring.adapter.dto.request.CreateDiaryRequestDTO
-import com.example.ggyunispring.adapter.dto.response.CreateDiaryResponseDTO
-import com.example.ggyunispring.adapter.dto.response.DiaryResponseDTO
+import com.example.ggyunispring.adapter.api.diary.CreateDiaryRequestDTO
+import com.example.ggyunispring.adapter.api.diary.CreateDiaryResponseDTO
+import com.example.ggyunispring.adapter.api.diary.DiaryResponseDTO
 import org.modelmapper.ModelMapper
 import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.security.core.userdetails.UserDetails
